# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TaleMotion Backend Foundation API
  version: 0.1.0
  description: PostgreSQL-backed project, scene, and job foundation.
servers:
  - url: http://localhost:8000/api/v1
security:
  - cookieAuth: []
paths:
...
  /auth/register:
    post:
      security: []
      tags: [Authentication]
      summary: Register a user and establish a cookie session
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RegisterRequest"}
      responses:
        "201":
          description: Registered user; session and CSRF cookies are set
          content:
            application/json:
              schema: {$ref: "#/components/schemas/UserResponse"}
        "409": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
  /auth/login:
    post:
      security: []
      tags: [Authentication]
      summary: Establish a cookie session
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LoginRequest"}
      responses:
        "200":
          description: Authenticated user; session and CSRF cookies are set
          content:
            application/json:
              schema: {$ref: "#/components/schemas/UserResponse"}
        "401": {$ref: "#/components/responses/Error"}
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke the current session
      parameters:
        - {$ref: "#/components/parameters/CsrfToken"}
      responses:
        "204": {description: Session revoked and cookies cleared}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get the current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: {$ref: "#/components/schemas/UserResponse"}
        "401": {$ref: "#/components/responses/Error"}
...
  /jobs:
    get:
      tags: [Jobs]
      summary: List persisted project jobs
      parameters:
        - {name: project_id, in: query, required: true, schema: {type: string}}
        - {name: active_only, in: query, schema: {type: boolean, default: false}}
      responses:
        "200":
          description: Project jobs, newest first
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GenerationJobListResponse"}
        "404": {$ref: "#/components/responses/Error"}
  /jobs/{job_id}:
    get:
      tags: [Jobs]
      summary: Inspect a persisted job
      parameters:
        - {$ref: "#/components/parameters/JobId"}
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GenerationJobResponse"}
        "404": {$ref: "#/components/responses/Error"}
  /jobs/{job_id}/cancel:
    post:
      tags: [Jobs]
      summary: Request job cancellation
      parameters:
        - {$ref: "#/components/parameters/JobId"}
      responses:
        "200":
          description: Job marked cancel_requested
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GenerationJobResponse"}
        "409": {$ref: "#/components/responses/Error"}
  /jobs/{job_id}/retry:
    post:
      tags: [Jobs]
      summary: Retry failed or cancelled generation work
      description: Creates a new attempt and preserves completed assets and prior job history.
      parameters:
        - {$ref: "#/components/parameters/JobId"}
      responses:
        "200":
          description: New queued retry job
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GenerationJobResponse"}
        "409": {$ref: "#/components/responses/Error"}
  /assets/{asset_id}:
    get:
      tags: [Assets]
      summary: Get persisted asset metadata
      parameters:
        - {name: asset_id, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Asset metadata
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AssetResponse"}
        "404": {$ref: "#/components/responses/Error"}
  /assets/{asset_id}/preview-url:
    post:
      tags: [Assets]
      summary: Create a short-lived signed B2 preview URL
      parameters:
        - {name: asset_id, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Signed preview
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SignedPreviewUrlResponse"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "502": {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: talemotion_session
  parameters:
...
    JobId: {name: job_id, in: path, required: true, schema: {type: string}}
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string}
  responses:
    Error:
      description: Consistent error envelope
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email: {type: string, format: email}
        password: {type: string, minLength: 12, maxLength: 128}
        name: {type: string, minLength: 1, maxLength: 120}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string, minLength: 1, maxLength: 128}
    UserResponse:
      type: object
      required: [id, email, name, created_at, updated_at]
      properties:
        id: {type: string}
        email: {type: string, format: email}
        name: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    JobType:
      type: string
      enum: [storyboard, scene_generation, scene_regeneration, project_generation, render]
    JobStatus:
      type: string
      enum: [queued, running, completed, failed, cancel_requested, cancelled]
...
    GenerationJobResponse:
      type: object
      required: [id, project_id, type, status, progress, input_payload, retry_count, max_retries, created_at, updated_at, children]
      properties:
        id: {type: string}
        project_id: {type: string}
        scene_id: {type: [string, "null"]}
        parent_job_id: {type: [string, "null"]}
        type: {$ref: "#/components/schemas/JobType"}
        status: {$ref: "#/components/schemas/JobStatus"}
        progress: {type: integer, minimum: 0, maximum: 100}
        current_stage: {type: [string, "null"]}
        input_payload: {type: object}
        result_payload: {type: [object, "null"]}
        error_code: {type: [string, "null"]}
        error_message: {type: [string, "null"]}
        retry_count: {type: integer}
        max_retries: {type: integer}
        cancel_requested_at: {type: [string, "null"], format: date-time}
        created_at: {type: string, format: date-time}
        started_at: {type: [string, "null"], format: date-time}
        completed_at: {type: [string, "null"], format: date-time}
        updated_at: {type: string, format: date-time}
        children:
          type: array
          items: {$ref: "#/components/schemas/GenerationJobChildResponse"}
    GenerationJobListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/GenerationJobResponse"}
    GenerationJobChildResponse:
      type: object
      required: [id, scene_id, status, progress, result_asset_id, error_code, error_message]
      properties:
        id: {type: string}
        scene_id: {type: [string, "null"]}
        status: {$ref: "#/components/schemas/JobStatus"}
        progress: {type: integer, minimum: 0, maximum: 100}
        result_asset_id: {type: [string, "null"]}
        error_code: {type: [string, "null"]}
        error_message: {type: [string, "null"]}
    AssetResponse:
      type: object
      required: [id, project_id, type, status, version, generation_parameters, created_at, updated_at]
      properties:
        id: {type: string}
        project_id: {type: string}
        scene_id: {type: [string, "null"]}
        parent_asset_id: {type: [string, "null"]}
        type: {type: string, enum: [image, video, audio, subtitle, manifest, thumbnail, final_video]}
        status: {type: string, enum: [pending, available, failed, archived]}
        version: {type: integer, minimum: 1}
        provider: {type: [string, "null"]}
        model_name: {type: [string, "null"]}
        prompt: {type: [string, "null"]}
        generation_parameters: {type: object}
        storage_bucket: {type: [string, "null"]}
        storage_object_key: {type: [string, "null"]}
        mime_type: {type: [string, "null"]}
        file_size_bytes: {type: [integer, "null"]}
        sha256: {type: [string, "null"]}
        provenance_object_key: {type: [string, "null"]}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        archived_at: {type: [string, "null"], format: date-time}
    SignedPreviewUrlResponse:
      type: object
      required: [url, expires_at]
      properties:
        url: {type: string, format: uri}
        expires_at: {type: string, format: date-time}
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            details: {type: object}
            request_id: {type: string}